refactor: update FellowshipReferendaPostList to use titleCount and titleExtra props, #6311
