+++ packages/next-common/pages/fellowship/index.js
@@ -9,10 +9,13 @@ import {
 import FellowshipReferendaPostList from "next-common/components/postList/fellowshipReferendaPostList";
 import normalizeFellowshipReferendaListItem from "next-common/utils/gov2/list/normalizeFellowshipReferendaListItem";
 import { fetchOpenGovTracksProps } from "next-common/services/serverSide";
+import NewFellowshipProposalButton from "next-common/components/summary/newFellowshipProposalButton";
 import CollectivesProvider from "next-common/context/collectives/collectives";
+import UnVotedOnlyOption from "next-common/components/referenda/unVotedOnlyOption";
 import useMyUnVotedCollectiveReferenda from "next-common/hooks/referenda/useMyUnVotedCollectiveReferenda";
 import { useEffect, useMemo, useState } from "react";
 import { usePageProps } from "next-common/context/page";
+import useRealAddress from "next-common/utils/hooks/useRealAddress";
 import {
   UnVotedOnlyProvider,
   useUnVotedOnlyContext,
@@ -62,6 +65,8 @@ function WithFilterPostList({
   pagination,
 }) {
   const { fellowshipTracks } = usePageProps();
+  const address = useRealAddress();
+  const { unVotedOnly, setUnVotedOnly } = useUnVotedOnlyContext();
 
   const items = (posts || []).map((item) =>
     normalizeFellowshipReferendaListItem(item, fellowshipTracks),
@@ -69,11 +74,22 @@ function WithFilterPostList({
 
   return (
     <FellowshipReferendaPostList
-      isUnVotedOnlyLoading={isUnVotedOnlyLoading}
       items={items}
       pagination={pagination}
-      total={total}
-      hasUnVotedOption
+      titleCount={total}
+      titleExtra={
+        <div className="flex gap-[12px] items-center">
+          {address && (
+            <UnVotedOnlyOption
+              tooltip="Only referenda I can but haven't voted"
+              isLoading={isUnVotedOnlyLoading}
+              isOn={unVotedOnly}
+              setIsOn={setUnVotedOnly}
+            />
+          )}
+          <NewFellowshipProposalButton />
+        </div>
+      }
     />
   );
 }
